changed county_poverty_population to integer

+++ api/app.py
@@ -76,7 +76,7 @@ class County(db.Model):
     state = db.Column(db.String(40), unique=False, nullable=False)
 
     county_poverty_percentage = db.Column(db.Float)
-    county_poverty_population = db.Column(db.Float)
+    county_poverty_population = db.Column(db.Integer)
 
     __table_args__ = (db.Index('county_state_index', "name", "state"), )
